backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pdfplumber>=0.10.3
apscheduler>=3.10.4
redis>=5.0.1
google-generativeai>=0.3.1
firebase-admin>=6.4.0
